Fix double-disconnect assertion when creating unusable account
When creating a new account that can not be connected to, a timeout
error is triggered after 30 seconds. If the app was put into background,
a disconnect and receiveQueue freeze could happen before the timeout
handler tries to disconnect again.
Prviosuly this triggered an assertion, now we try to handle that case
more gracefully while not introducing new race conditions.